requirements.txt
pandas
numpy
streamlit
sqlalchemy
python-dotenv
